hotfix: fix the fixtures and models
Explanation: the issue was about Deserialization on the fixtures data, the way to solve this bug was dump the existing data to look the greate format to make the fixtures. On the other hand was required refactor the create_user method on Custom User model to separate the logic between create a user and create a super user because for fix the bug I had to use the command createsuperuser and for that the custom user manager has to have the create_superuser method isolated. Dor more information about that look this site: https://docs.djangoproject.com/en/4.2/topics/db/fixtures/